+++ src/animation/mergeanimation.cpp
@@ -6,12 +6,13 @@ namespace PixelMaestro {
 	}
 
 	void MergeAnimation::update() {
+		Point center = get_center();
 		if (orientation_ == Vertical) {
 			for (uint16_t x = 0; x < section_->get_dimensions()->x; x++) {
 				count_ = 0;
 
 				// Note: COLUMN MUST BE A SIGNED INT IN ORDER TO ACCESS INDEX 0.
-				for (int32_t y = center_.y; y >= 0; y--) {
+				for (int32_t y = center.y; y >= 0; y--) {
 					section_->set_one(x, y, get_color_at_index(count_ + cycle_index_));
 					count_++;
 				}
@@ -23,14 +24,14 @@ namespace PixelMaestro {
 				uint8_t offset = 0;
 				if (section_->get_dimensions()->size() % 2 != 0) {
 					offset += 1;
-					section_->set_one(x, center_.y + offset, get_color_at_index(cycle_index_));
+					section_->set_one(x, center.y + offset, get_color_at_index(cycle_index_));
 				}
 
 				offset += 1;
 
 				// Go from the center to the last
 				count_ = 0;
-				for (uint16_t y = center_.y + offset; y < section_->get_dimensions()->y; y++) {
+				for (uint16_t y = center.y + offset; y < section_->get_dimensions()->y; y++) {
 					section_->set_one(x, y, get_color_at_index(count_ + cycle_index_));
 					count_++;
 				}
@@ -41,7 +42,7 @@ namespace PixelMaestro {
 				count_ = 0;
 
 				// Note: ROW MUST BE A SIGNED INT IN ORDER TO ACCESS INDEX 0.
-				for (int32_t x = center_.x; x >= 0; x--) {
+				for (int32_t x = center.x; x >= 0; x--) {
 					section_->set_one(x, y, get_color_at_index(count_ + cycle_index_));
 					count_++;
 				}
@@ -53,14 +54,14 @@ namespace PixelMaestro {
 				uint8_t offset = 0;
 				if (section_->get_dimensions()->size() % 2 != 0) {
 					offset += 1;
-					section_->set_one(center_.x + offset, y, get_color_at_index(cycle_index_));
+					section_->set_one(center.x + offset, y, get_color_at_index(cycle_index_));
 				}
 
 				offset += 1;
 
 				// Go from the center to the last
 				count_ = 0;
-				for (uint16_t x = center_.x + offset; x < section_->get_dimensions()->x; x++) {
+				for (uint16_t x = center.x + offset; x < section_->get_dimensions()->x; x++) {
 					section_->set_one(x, y, get_color_at_index(count_ + cycle_index_));
 					count_++;
 				}