+++ src/cue/animationcuehandler.h
@@ -10,8 +10,6 @@ namespace PixelMaestro {
 	class AnimationCueHandler : public CueHandler {
 		public:
 			enum Action : uint8_t {
-				ResetCenter,
-				SetCenter,
 				SetColors,
 				SetCycleIndex,
 				SetFade,
@@ -42,9 +40,6 @@ namespace PixelMaestro {
 			uint8_t* set_sparkle_options(uint8_t section_num, uint8_t layer_num, uint8_t threshold);
 
 			// General Animation calls
-			uint8_t* reset_center(uint8_t section_num, uint8_t layer_num);
-			uint8_t* set_center(uint8_t section_num, uint8_t layer_num);
-			uint8_t* set_center(uint8_t section_num, uint8_t layer_num, uint16_t center_x, uint16_t center_y);
 			uint8_t* set_colors(uint8_t section_num, uint8_t layer_num, Colors::RGB* colors, uint8_t num_colors, bool delete_old_colors = true);
 			uint8_t* set_cycle_index(uint8_t section_num, uint8_t layer_num, uint8_t cycle_index);
 			uint8_t* set_fade(uint8_t section_num, uint8_t layer_num, bool fade);