+++ src/cue/sectioncuehandler.h
@@ -15,7 +15,9 @@ namespace PixelMaestro {
 				SetAnimation,
 				SetCanvas,
 				SetDimensions,
-				SetLayer
+				SetLayer,
+				SetOffset,
+				SetScroll
 			};
 
 			enum Byte : uint8_t {
@@ -34,6 +36,8 @@ namespace PixelMaestro {
 			uint8_t* set_canvas(uint8_t section_num, uint8_t layer_num, CanvasType::Type canvas_type, uint16_t num_frames = 1);
 			uint8_t* set_dimensions(uint8_t section_num, uint8_t layer_num, uint16_t x, uint16_t y);
 			uint8_t* set_layer(uint8_t section_num, uint8_t layer_num, Colors::MixMode mix_mode, uint8_t alpha);
+			uint8_t* set_offset(uint8_t section_num, uint8_t layer_num, int16_t x, int16_t y);
+			uint8_t* set_scroll(uint8_t section_num, uint8_t layer_num, int16_t x, int16_t y);
 			void run(uint8_t* cue);
 	};
 }